requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
Fabric3==1.14.post1
mysqlclient==2.1.1
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
six==1.16.0
sqlparse==0.4.3
